pyproject.toml
[tool.poetry]
name = "wqunits"
version = "0.1.0"
description = ""
authors = ["Mateusz <zawadzkimat@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pint = "^0.24.3"
pyyaml = "^6.0.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
